+++ tests/testthat/test-drop_diseasystore.R
@@ -19,7 +19,11 @@ test_that("drop_diseasystore can delete entire default schema", {
 
 
     for (id in ids) {
-      SCDB::create_table(mtcars, conn, id, temporary = FALSE)
+      if (packageVersion("SCDB") <= "0.3") {
+        DBI::dbWriteTable(conn, id, mtcars, temporary = FALSE, overwrite = TRUE)
+      } else {
+        SCDB::create_table(mtcars, conn, id, temporary = FALSE)
+      }
     }
 
     # Try to delete the entire `target_schema_1` store
@@ -27,26 +31,21 @@ test_that("drop_diseasystore can delete entire default schema", {
     expect_identical(diseasyoption("target_schema"), target_schema_1)
     drop_diseasystore(conn = conn)
 
-    expect_false(SCDB::table_exists(conn, paste(target_schema_1, "logs",     sep = ".")))
-    expect_false(SCDB::table_exists(conn, paste(target_schema_1, "mtcars_1", sep = ".")))
-    expect_false(SCDB::table_exists(conn, paste(target_schema_1, "mtcars_2", sep = ".")))
+    expect_false(DBI::dbExistsTable(conn, logs_id))
+    expect_false(DBI::dbExistsTable(conn, ids[[1]]))
+    expect_false(DBI::dbExistsTable(conn, ids[[2]]))
 
-    expect_true(SCDB::table_exists(conn, "mtcars_1"))
-    expect_true(SCDB::table_exists(conn, paste(target_schema_2, "mtcars_1", sep = ".")))
+    expect_true(DBI::dbExistsTable(conn, ids[[3]]))
+    expect_true(DBI::dbExistsTable(conn, ids[[4]]))
 
 
     # Make sure all tables have been removed
-    c(paste(target_schema_1, "logs",     sep = "."),
-      paste(target_schema_1, "mtcars_1", sep = "."),
-      paste(target_schema_1, "mtcars_2", sep = "."),
-      "mtcars_1",
-      paste(target_schema_2, "mtcars_1", sep = ".")
-    ) |>
+    c(SCDB::id(paste(target_schema_1, "logs", sep = "."), conn), ids) |>
       purrr::walk(~ {
-        if (SCDB::table_exists(conn, .)) {
-          DBI::dbRemoveTable(conn, SCDB::id(., conn))
+        if (DBI::dbExistsTable(conn, .)) {
+          DBI::dbRemoveTable(conn, .)
         }
-        expect_false(SCDB::table_exists(conn, .))
+        expect_false(DBI::dbExistsTable(conn, .))
       })
 
     DBI::dbDisconnect(conn)
@@ -73,7 +72,11 @@ test_that("drop_diseasystore can delete single table in default schema", {
       purrr::map(~ SCDB::id(., conn))
 
     for (id in ids) {
-      SCDB::create_table(mtcars, conn, id, temporary = FALSE)
+      if (packageVersion("SCDB") <= "0.3") {
+        DBI::dbWriteTable(conn, id, mtcars, temporary = FALSE, overwrite = TRUE)
+      } else {
+        SCDB::create_table(mtcars, conn, id, temporary = FALSE)
+      }
     }
 
     # Try to delete only mtcars_1 within the diseasystore
@@ -81,37 +84,32 @@ test_that("drop_diseasystore can delete single table in default schema", {
     expect_identical(diseasyoption("target_schema"), target_schema_1)
     drop_diseasystore(pattern = "mtcars_1", conn = conn)
 
-    expect_true(SCDB::table_exists(conn, paste(target_schema_1, "logs", sep = ".")))
-    expect_false(SCDB::table_exists(conn, paste(target_schema_1, "mtcars_1", sep = ".")))
-    expect_true(SCDB::table_exists(conn, paste(target_schema_1, "mtcars_2", sep = ".")))
+    expect_true(DBI::dbExistsTable(conn, logs_id))
+    expect_false(DBI::dbExistsTable(conn, ids[[1]]))
+    expect_true(DBI::dbExistsTable(conn, ids[[2]]))
 
-    expect_true(SCDB::table_exists(conn, "mtcars_1"))
-    expect_true(SCDB::table_exists(conn, paste(target_schema_2, "mtcars_1", sep = ".")))
+    expect_true(DBI::dbExistsTable(conn, ids[[3]]))
+    expect_true(DBI::dbExistsTable(conn, ids[[4]]))
 
     # Try to delete only mtcars_2 within the diseasystore
     # But first, verify that the testing target_schema has been set
     expect_identical(diseasyoption("target_schema"), target_schema_1)
     drop_diseasystore(pattern = "mtcars_2", conn = conn)
 
-    expect_true(SCDB::table_exists(conn, paste(target_schema_1, "logs", sep = ".")))
-    expect_false(SCDB::table_exists(conn, paste(target_schema_1, "mtcars_1", sep = ".")))
-    expect_false(SCDB::table_exists(conn, paste(target_schema_1, "mtcars_2", sep = ".")))
+    expect_true(DBI::dbExistsTable(conn, logs_id))
+    expect_false(DBI::dbExistsTable(conn, ids[[1]]))
+    expect_false(DBI::dbExistsTable(conn, ids[[2]]))
 
-    expect_true(SCDB::table_exists(conn, "mtcars_1"))
-    expect_true(SCDB::table_exists(conn, paste(target_schema_2, "mtcars_1", sep = ".")))
+    expect_true(DBI::dbExistsTable(conn, ids[[3]]))
+    expect_true(DBI::dbExistsTable(conn, ids[[4]]))
 
     # Make sure all tables have been removed
-    c(paste(target_schema_1, "logs",     sep = "."),
-      paste(target_schema_1, "mtcars_1", sep = "."),
-      paste(target_schema_1, "mtcars_2", sep = "."),
-      "mtcars_1",
-      paste(target_schema_2, "mtcars_1", sep = ".")
-    ) |>
+    c(SCDB::id(paste(target_schema_1, "logs", sep = "."), conn), ids) |>
       purrr::walk(~ {
-        if (SCDB::table_exists(conn, .)) {
-          DBI::dbRemoveTable(conn, SCDB::id(., conn))
+        if (DBI::dbExistsTable(conn, .)) {
+          DBI::dbRemoveTable(conn, .)
         }
-        expect_false(SCDB::table_exists(conn, .))
+        expect_false(DBI::dbExistsTable(conn, .))
       })
 
     DBI::dbDisconnect(conn)